Fix crate build with clang/LLVM 16
The dependent carate `blkid-sys` has already a fix in `master` but it
was not released yet. And we are using this custom fork with the custom
branch. So, I have to update it with the fix.

+++ src/prober.rs
@@ -426,7 +426,7 @@ impl Prober {
 
     /// Returns name of a supported partition.
     #[cfg(blkid = "2.30")]
-    pub fn partitions_get_name(idx: u64) -> BlkIdResult<String> {
+    pub fn partitions_get_name(idx: usize) -> BlkIdResult<String> {
         let mut name: *const ::libc::c_char = ptr::null();
         unsafe { c_result(blkid_partitions_get_name(idx, &mut name)) }?;
         let name = unsafe { CStr::from_ptr(name).to_str()?.to_owned() };
